Inventario-Django/requirements.txt
asgiref==3.9.1
attrs==25.4.0
black==25.11.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
Django==5.2.6
django-cors-headers==4.9.0
django-model-utils==5.0.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
drf-yasg==1.21.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy_extensions==1.1.0
mysqlclient==2.2.7
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
PyJWT==2.10.1
pytest==8.4.2
pytest-django==4.11.1
pytokens==0.3.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.37.0
reportlab==4.4.4
rpds-py==0.28.0
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.2.0
whitenoise==6.11.0